[SDLMESS] Fix input on big-endian systems.
git-svn-id: svn://messdev.no-ip.org/mess@724 9849079f-6229-0410-a07f-84406ce305ba

+++ src/osd/sdl/input.c
@@ -1770,7 +1770,7 @@ static void generic_device_reset(device_info *devinfo)
 
 static INT32 generic_button_get_state(void *device_internal, void *item_internal)
 {
-	UINT8 *itemdata = item_internal;
+	INT32 *itemdata = item_internal;
 	
 	// return the current state
 	return *itemdata >> 7;
